fix: inline compression logic in Appwrite functions
